requirements.txt
async-generator==1.10
attrs==22.2.0
beautifulsoup4==4.11.2
blinker==1.5
Brotli==1.0.9
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
colorama==0.4.6
cryptography==39.0.1
exceptiongroup==1.1.0
fake-useragent==1.1.1
google==3.0.0
google-api-core==2.11.0
google-api-python-client==2.81.0
google-auth==2.16.2
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.58.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httplib2==0.21.0
hyperframe==6.0.1
idna==3.4
kaitaistruct==0.10
MouseInfo==0.1.3
oauthlib==3.2.2
outcome==1.2.0
packaging==23.0
Pillow==9.4.0
protobuf==4.22.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAutoGUI==0.9.53
pycparser==2.21
pydivert==2.1.0
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyOpenSSL==23.0.0
pyparsing==3.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.28
PySocks==1.7.1
python-dotenv==0.21.1
pytweening==1.0.4
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
selenium==4.8.0
selenium-wire==5.1.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
tqdm==4.64.1
trio==0.22.0
trio-websocket==0.9.2
undetected-chromedriver==3.4.4
uritemplate==4.1.1
urllib3==1.26.14
webdriver-manager==3.8.5
websockets==10.4
wsproto==1.2.0
zstandard==0.19.0